+++ src/modules/parachute.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "modules/prelude.h"
+
+namespace hff {
+
+extern bool g_parachuteFlight;
+
+bool InstallParachuteFlightFix();
+void UpdateParachuteScript();
+
+} // namespace hff